Validations

+++ Inbursa.Server/Models/Cuenta.cs
@@ -5,8 +5,8 @@ namespace Inbursa.Server.Models;
 public class Cuenta
 {
     public String? Usuario { get; set; }
-    public float Saldo { get; set; }
+    public double Saldo { get; set; }
     [Key]
     public int Num_Cuenta { get; set; }
-   public bool Prestamo_Activo { get; set; }
+    public bool Prestamo_Activo { get; set; }
 }
